[refactoring] split SSDCaller and TestScenario
