--- src/main/java/shell/TestScenario.java
@@ -1,14 +1,0 @@
-package shell;
-
-import java.io.IOException;
-import java.util.Random;
-
-public class TestScenario {
-    TestShell testShell;
-    Random rand;
-
-    public TestScenario(TestShell testShell, Random rand) {
-        this.testShell = testShell;
-        this.rand = rand;
-    }
-}
