+++ src/main/java/shell/command/Command.java
@@ -1,4 +1,4 @@
-package shell;
+package shell.command;
 
 public interface Command {
     void execute(String[] args);
