+++ src/main/java/shell/util/Utility.java
@@ -1,0 +1,4 @@
+package shell.util;
+
+public class Utility {
+}
